Removes updateConfig and planStrategy front configuration. Moves shared and reuseable classes out of backup package.
