--- cassandra-commons/src/main/java/com/mesosphere/dcos/cassandra/common/backup/ClusterTaskContext.java
@@ -1,4 +1,0 @@
-package com.mesosphere.dcos.cassandra.common.backup;
-
-public interface ClusterTaskContext {
-}
