+++ cassandra-commons/src/main/java/com/mesosphere/dcos/cassandra/common/tasks/ClusterTaskContext.java
@@ -1,0 +1,4 @@
+package com.mesosphere.dcos.cassandra.common.tasks;
+
+public interface ClusterTaskContext {
+}
