+++ cassandra-scheduler/src/main/java/com/mesosphere/dcos/cassandra/scheduler/config/ConfigurationManager.java
@@ -1,11 +1,12 @@
 package com.mesosphere.dcos.cassandra.scheduler.config;
 
 
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
-import com.mesosphere.dcos.cassandra.common.backup.BackupContext;
-import com.mesosphere.dcos.cassandra.common.backup.RestoreContext;
+import com.mesosphere.dcos.cassandra.common.tasks.backup.BackupContext;
+import com.mesosphere.dcos.cassandra.common.tasks.backup.RestoreContext;
 import com.mesosphere.dcos.cassandra.common.config.CassandraApplicationConfig;
 import com.mesosphere.dcos.cassandra.common.config.CassandraConfig;
 import com.mesosphere.dcos.cassandra.common.config.ClusterTaskConfig;
@@ -20,9 +21,7 @@ import org.apache.mesos.Protos;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Arrays;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
 
 
 public class ConfigurationManager implements Managed {
@@ -40,57 +39,50 @@ public class ConfigurationManager implements Managed {
     private volatile ExecutorConfig executorConfig;
     private volatile int servers;
     private volatile int seeds;
-    private volatile boolean updateConfig;
     private final String placementStrategy;
-    private final String planStrategy;
     private final String seedsUrl;
+    private final List<String> errors = new ArrayList<>();
 
     private void reconcileConfiguration() throws PersistenceException {
         LOGGER.info("Reconciling remote and persisted configuration");
 
         Optional<Integer> serversOption = serversRef.load();
 
-        if (updateConfig) {
-            LOGGER.info("Configuration update requested");
 
-            if (serversOption.isPresent()) {
-                int servers = serversOption.get();
-                if (this.servers < servers) {
-                    String error = String.format("The number of configured " +
-                                    "servers (%d) is less than the current " +
-                                    "number of configured servers (%d). Reduce the " +
-                                    "number of servers by removing them from the cluster",
-                            this.servers,
-                            servers);
-                    LOGGER.error(error);
-                    throw new IllegalStateException(error);
-                }
+        LOGGER.info("Configuration update requested");
 
-            }
-
-            if (seeds > servers) {
+        if (serversOption.isPresent()) {
+            int servers = serversOption.get();
+            if (this.servers < servers) {
                 String error = String.format("The number of configured " +
-                                "seeds (%d) is greater than the current number " +
-                                "of configured servers (%d). Reduce the " +
-                                "number of seeds or increase the number of servers",
-                        seeds,
+                                "servers (%d) is less than the current " +
+                                "number of configured servers (%d). Reduce the " +
+                                "number of servers by removing them from the cluster",
+                        this.servers,
                         servers);
                 LOGGER.error(error);
-                throw new IllegalStateException(error);
+                errors.add(error);
             }
 
-            serversRef.store(servers);
-            seedsRef.store(seeds);
-            cassandraRef.store(cassandraConfig);
-            executorRef.store(executorConfig);
-
-        } else {
-            LOGGER.info("Using persisted configuration");
-            servers = serversRef.putIfAbsent(servers);
-            seeds = seedsRef.putIfAbsent(seeds);
-            cassandraConfig = cassandraRef.putIfAbsent(cassandraConfig);
-            executorConfig = executorRef.putIfAbsent(executorConfig);
         }
+
+        if (seeds > servers) {
+            String error = String.format("The number of configured " +
+                            "seeds (%d) is greater than the current number " +
+                            "of configured servers (%d). Reduce the " +
+                            "number of seeds or increase the number of servers",
+                    seeds,
+                    servers);
+            LOGGER.error(error);
+            errors.add(error);
+        }
+
+        serversRef.store(servers);
+        seedsRef.store(seeds);
+        cassandraRef.store(cassandraConfig);
+        executorRef.store(executorConfig);
+
+
     }
 
     @Inject
@@ -100,9 +92,7 @@ public class ConfigurationManager implements Managed {
             @Named("ConfiguredExecutorConfig") ExecutorConfig executorConfig,
             @Named("ConfiguredServers") int servers,
             @Named("ConfiguredSeeds") int seeds,
-            @Named("ConfiguredUpdateConfig") boolean updateConfig,
             @Named("ConfiguredPlacementStrategy") String placementStrategy,
-            @Named("ConfiguredPlanStrategy") String planStrategy,
             @Named("SeedsUrl") String seedsUrl,
             PersistenceFactory persistenceFactory,
             Serializer<CassandraConfig> cassandraConfigSerializer,
@@ -130,9 +120,7 @@ public class ConfigurationManager implements Managed {
         this.executorConfig = executorConfig;
         this.servers = servers;
         this.seeds = seeds;
-        this.updateConfig = updateConfig;
         this.placementStrategy = placementStrategy;
-        this.planStrategy = planStrategy;
         this.seedsUrl = seedsUrl;
 
         try {
@@ -162,10 +150,6 @@ public class ConfigurationManager implements Managed {
 
     public String getPlacementStrategy() {
         return placementStrategy;
-    }
-
-    public String getPlanStrategy() {
-        return planStrategy;
     }
 
     public void setCassandraConfig(final CassandraConfig cassandraConfig)
@@ -263,7 +247,8 @@ public class ConfigurationManager implements Managed {
                         setApplication(cassandraConfig.getApplication()
                                 .toBuilder().setSeedProvider(
                                         CassandraApplicationConfig
-                                                .createDcosSeedProvider(seedsUrl))
+                                                .createDcosSeedProvider(
+                                                        seedsUrl))
                                 .build())
                         .build(),
                 CassandraDaemonStatus.create(Protos.TaskState.TASK_STAGING,
@@ -414,30 +399,30 @@ public class ConfigurationManager implements Managed {
 
     public boolean hasCurrentExecutorConfig(
             final CassandraTaskExecutor executor) {
-       return executor.getCommand().equals(executorConfig
+        return executor.getCommand().equals(executorConfig
                 .getCommand()) &&
                 Double.compare(executor.getCpus(),
                         executorConfig.getCpus()) == 0 &&
                 executor.getDiskMb() == executorConfig.getDiskMb() &&
                 Double.compare(executor.getCpus(),
                         executorConfig.getCpus()) == 0 &&
-               executor.getDiskMb() ==
+                executor.getDiskMb() ==
                         executorConfig.getDiskMb() &&
-               executor.getMemoryMb() ==
+                executor.getMemoryMb() ==
                         executorConfig.getMemoryMb() &&
-               executor.getUriStrings().containsAll(
+                executor.getUriStrings().containsAll(
                         Arrays.asList(
                                 executorConfig.getCassandraLocationString(),
                                 executorConfig.getExecutorLocationString(),
                                 executorConfig.getJreLocationString()
                         )
                 ) &&
-               executor.getHeapMb() == executorConfig.getHeapMb();
+                executor.getHeapMb() == executorConfig.getHeapMb();
     }
 
     public boolean hasCurrentConfig(final CassandraDaemonTask task) {
 
-        return  hasCurrentExecutorConfig(task.getExecutor()) &&
+        return hasCurrentExecutorConfig(task.getExecutor()) &&
                 task.getConfig().equals(cassandraConfig);
 
     }
@@ -449,7 +434,7 @@ public class ConfigurationManager implements Managed {
                 id,
                 task.getSlaveId(),
                 task.getHostname(),
-                updateExecutor(task,id),
+                updateExecutor(task, id),
                 task.getName(),
                 task.getRole(),
                 task.getPrincipal(),
@@ -476,6 +461,10 @@ public class ConfigurationManager implements Managed {
     public CassandraTask updateId(CassandraTask task) {
         return task.updateId(
                 task.getName() + "_" + UUID.randomUUID().toString());
+    }
+
+    public List<String> getErrors() {
+        return ImmutableList.copyOf(errors);
     }
 
     @Override
