+++ cassandra-scheduler/src/main/java/com/mesosphere/dcos/cassandra/scheduler/offer/ClusterTaskPlacementStrategy.java
@@ -1,8 +1,6 @@
-package com.mesosphere.dcos.cassandra.scheduler.backup;
+package com.mesosphere.dcos.cassandra.scheduler.offer;
 
 import com.google.common.collect.Lists;
-import com.mesosphere.dcos.cassandra.common.tasks.CassandraDaemonTask;
-import com.mesosphere.dcos.cassandra.common.util.TaskUtils;
 import com.mesosphere.dcos.cassandra.scheduler.tasks.CassandraTasks;
 import org.apache.mesos.Protos;
 import org.apache.mesos.offer.PlacementStrategy;
@@ -12,7 +10,6 @@ import org.slf4j.LoggerFactory;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 
 public class ClusterTaskPlacementStrategy implements PlacementStrategy {
     private static final Logger LOGGER = LoggerFactory.getLogger(
