+++ cassandra-scheduler/src/main/java/com/mesosphere/dcos/cassandra/scheduler/resources/RestoreResource.java
@@ -2,10 +2,8 @@ package com.mesosphere.dcos.cassandra.scheduler.resources;
 
 import com.codahale.metrics.annotation.Timed;
 import com.google.inject.Inject;
-import com.mesosphere.dcos.cassandra.common.backup.RestoreContext;
+import com.mesosphere.dcos.cassandra.common.tasks.backup.RestoreContext;
 import com.mesosphere.dcos.cassandra.scheduler.backup.RestoreManager;
-import org.apache.mesos.scheduler.plan.StageManager;
-import org.apache.mesos.scheduler.plan.api.StageInfo;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
