+++ cassandra-scheduler/src/test/java/com/mesosphere/dcos/cassandra/scheduler/resources/ConfigurationResourceTest.java
@@ -76,9 +76,7 @@ public class ConfigurationResourceTest {
                 config.getExecutorConfig(),
                 config.getServers(),
                 config.getSeeds(),
-                false,
                 "NODE",
-                "INSTALL",
                 config.getSeedsUrl(),
                 persistence,
                 CassandraConfig.JSON_SERIALIZER,
