+++ cassandra-scheduler/src/test/java/com/mesosphere/dcos/cassandra/scheduler/tasks/CassandraTasksTest.java
@@ -105,9 +105,7 @@ public class CassandraTasksTest {
                 config.getExecutorConfig(),
                 config.getServers(),
                 config.getSeeds(),
-                false,
                 "NODE",
-                "INSTALL",
                 config.getSeedsUrl(),
                 persistence,
                 CassandraConfig.JSON_SERIALIZER,
